rptest/wasm: Move WasmScript class into new file
(cherry picked from commit ca68a50ab18df88c12628997b44ac38d684eb3b0)
